--- a/fe/fe-core/src/main/java/com/starrocks/qe/Coordinator.java
+++ b/fe/fe-core/src/main/java/com/starrocks/qe/Coordinator.java
@@ -1125,18 +1125,8 @@ public class Coordinator {
                 if (dopAdaptionEnabled) {
                     int degreeOfParallelism = ConnectContext.get().getSessionVariable().getDegreeOfParallelism();
                     Preconditions.checkArgument(leftMostNode instanceof ExchangeNode);
-                    // If the maximum parallel child fragment send data to the current PlanFragment via UNPARTITIONED
-                    // DataStreamSink, then fragment instance parallelization is leveraged, otherwise, pipeline parallelization
-                    // is adopted.
-                    if (maxParallelismFragmentExecParams.fragment.isPartitioned()) {
-                        // fragment instance parallelization (numInstances=N, pipelineDop=1)
-                        maxParallelism = Math.min(hostSet.size() * degreeOfParallelism, maxParallelism);
-                        fragment.setPipelineDop(1);
-                    } else {
-                        // pipeline parallelization (numInstances=|hostSet|, pipelineDop=degreeOfParallelism)
-                        maxParallelism = hostSet.size();
-                        fragment.setPipelineDop(degreeOfParallelism);
-                    }
+                    maxParallelism = hostSet.size();
+                    fragment.setPipelineDop(degreeOfParallelism);
                 }
 
                 // AddAll() soft copy()
@@ -1975,10 +1965,11 @@ public class Coordinator {
                     SessionVariable sessionVariable = ConnectContext.get().getSessionVariable();
 
                     if (isEnablePipelineEngine) {
-                        boolean isPipeline = fragment.getPlanRoot().canUsePipeLine() && fragment.getSink().canUsePipeLine();
+                        boolean isPipeline =
+                                fragment.getPlanRoot().canUsePipeLine() && fragment.getSink().canUsePipeLine();
                         params.setIs_pipeline(isPipeline);
                         if (isPipeline) {
-                            queryOptions.setBatch_size(SessionVariable.PIPELINE_BATCH_SIZE); 
+                            queryOptions.setBatch_size(SessionVariable.PIPELINE_BATCH_SIZE);
                         }
                         params.setPipeline_dop(fragment.getPipelineDop());
                         params.setPer_scan_node_dop(instanceExecParam.perScanNodeDop);