--- a/gensrc/proto/internal_service.proto
+++ b/gensrc/proto/internal_service.proto
@@ -70,10 +70,12 @@ message PTransmitChunkParams {
 
     // Some statistics for the runing query
     optional PQueryStatistics query_statistics = 8;
-    optional bool use_pass_through = 9 [default = false] ;
+    optional bool use_pass_through = 9 [default = false];
 
-    // Pipeline level shuffle's id
-    optional int32 driver_sequence = 10;
+    // Whether enable pipeline level shuffle
+    optional bool is_pipeline_level_shuffle = 10 [default = false];
+    // Driver sequences of pipeline level shuffle
+    repeated int32 driver_sequences = 11;
 };
 
 message PTransmitDataResult {