## Background

For [issue-3013](https://github.com/StarRocks/starrocks/issues/3013)

When enable pipeline level shuffle, chunks will hash to N groups (N is the dop of dest pipeline). Each group will build the request independently. 

* When sending large amounts of data, the number of requests will not change significantly because each group will accumulate chunk until it exceeds a threshold before sending
* When sending small amounts of data, take the extreme case for example, assume the amount of data just fills one chunk, then each group will send one request, the requests to be sent changed from `1` to `N` which may lead to significant performance drop

## Solution

Put chunks of different driver_sequences in one request and add an auxiliary data structure `PTransmitChunkParams::driver_sequences`, indicating the `driver_sequence` to which the `i-th chunk` belongs

## Test

* 3be(64c/128G)
* 1fe(64c/128G)
* tpcds with scale factor=100

**Almost all the querys have performance improvements, and only a few are listed below.**

| TPCDS | Before | After |
|:--|:--|:--|
| Q12 | 0.299s | 0.19s |
| Q15 | 0.378s | 0.25s |
| Q20 | 0.25s | 0.16s |
| Q45 | 0.442s | 0.305s |