@liuyehcf 2022-01-22
run starrocks_be_unittest